Ошибка избранного обрабатывается в компоненте App
