+++ src/components/app/app.tsx
@@ -8,8 +8,20 @@ import NotFoundPage from '../../pages/not-found-page/not-found-page.tsx';
 import ProtectedRoute from '../protected-route/protected-route.tsx';
 import { AppRoute } from '../../const.ts';
 import MainPage from '../../pages/main-page/main-page.tsx';
+import { useAppDispatch, useAppSelector } from '../../hooks/index.tsx';
+import { getFetchingFavoriteError } from '../../store/favorites-page/selectors.ts';
+import { toast } from 'react-toastify';
+import { closeError } from '../../store/favorites-page/favorites-page.ts';
 
 export default function App() {
+
+  const hasError = useAppSelector(getFetchingFavoriteError);
+  const dispatch = useAppDispatch();
+
+  if (hasError) {
+    toast.warn('Something went wrong');
+    setTimeout(() => dispatch(closeError()), 5000);
+  }
 
   return (
     <HelmetProvider>
